% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsAlbers.r
\name{crsAlbers}
\alias{crsAlbers}
\title{Generate a custom coordinate reference system WKT2 strings}
\usage{
crsAlbers(x)
}
\arguments{
\item{x}{Either an object of class \code{SpatRaster}, \code{SpatVector}, or \code{sf}, \emph{or} a numeric vector with two values (longitude and latitude of the center of the projection), \emph{or} a two-column matrix/data frame with the centroid of the projection.}
}
\description{
These functions take as inpit either a spatial object or coordinate pair and generate centered on it a custom WKT2 (well-known text) coordinate reference system string. Projections include:
\itemize{
\item Albers conic equal-area
\item Lambert azimuthal equal-area
\item Vertical near-side (i.e., as the world appears from geosynchronous orbit)
}
}
\examples{
library(sf)

# Madagascar
data(mad0)

alb <- crsAlbers(mad0)
lamb <- crsLambert(mad0)
vert <- crsVertical(mad0)

madAlb <- st_transform(mad0, alb)
madLamb <- st_transform(mad0, lamb)
madVert <- st_transform(mad0, vert)

par(mfrow=c(2, 2))
plot(st_geometry(mad0), main='Unprojected (WGS84)')
plot(st_geometry(madAlb), main='Albers')
plot(st_geometry(madLamb), main='Lambert')
plot(st_geometry(madVert), main='Vertical')

# The effect is more noticable when plotting large areas,
# especially if they lie near the poles.
if (FALSE) {

library(geodata)
library(terra)

can <- gadm('CAN', level=0, path=tempdir()) # outline of Canada

alb <- crsAlbers(can)
lamb <- crsLambert(can)
vert <- crsVertical(can)

canAlb <- project(can, alb)
canLamb <- project(can, lamb)
canVert <- project(can, vert)

par(mfrow=c(2, 2))
plot(can, main='Unprojected (WGS84)')
plot(canAlb, main='Albers')
plot(canLamb, main='Lambert')
plot(canVert, main='Vertical')
	
}
}
\seealso{
\code{\link{crsGet}}, \code{\link{crsLambert}}, \code{\link{crsVNS}}
}
