% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsLambert.r
\name{crsLambert}
\alias{crsLambert}
\title{Generate a custom Lambert azimuthal equal-area projection WKT2 string}
\usage{
crsLambert(x)
}
\arguments{
\item{x}{Either an object of class \code{SpatRaster}, \code{SpatVector}, or \code{sf}, \emph{or} a numeric string with two values (longitude and latitude of the center of the projection), \emph{or} a two-column matrix/data frame with the centroid of the projection.}
}
\description{
This function generates a WKT2 (well-known text) coordinate reference string for a Lambert equal-area azimuthal projection of a geographic object or coordinate pair.
}
\examples{

library(sf)
data(mad0)
wkt2 <- crsLambert(mad0)
mad0Lamb <- st_transform(mad0, wkt2)

par(mfrow=c(1, 2))
plot(st_geometry(mad0), main='Unprojected (WGS84)')
plot(st_geometry(mad0Lamb), main='Lambert')

# The effect is more noticable when plotting large areas,
# especially near the poles.
if (FALSE) {

library(geodata)
library(terra)

can <- gadm('CAN', level=0, path=getwd()) # outline of Canada

wktVert <- crsVertical(can)
wktLamb <- crsLambert(can)

canVert <- project(can, wktVert)
canLamb <- project(can, wktLamb)

par(mfrow=c(1, 3))
plot(can, main='WGS84')
plot(canVert, main='Vertical')
plot(canLamb, main='Lambert')

}

}
\seealso{
\code{\link{crsGet}}
}
