% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsVertical.r
\name{crsVertical}
\alias{crsVertical}
\title{Generate a "Vertical Near-Side" projection WKT2 string}
\usage{
crsVertical(x, alt = 35800)
}
\arguments{
\item{x}{Either an object of class \code{SpatRaster}, \code{SpatVector}, or \code{sf}, \emph{or} a numeric string with two values (longitude and latitude of the center of the projection), \emph{or} a two-column matrix/data frame with the centroid of the projection.}

\item{alt}{Altitude in meters of the viewpoint in km. The default (35800 km) is geosynchronous orbit.}
}
\description{
This function generates a WKT2 (well-known text) coordinate reference string for a "vertical near-side" projection of a geographic object or coordinate pair. When plotted, such an object will appear as if it would from a satellite in geosyncronous orbit directly above the centroid of the region of interest (showing about 1/3 of the Earth).
}
\examples{

library(sf)
data(mad0)
wkt2 <- crsVertical(mad0)
mad0vert <- st_transform(mad0, wkt2)

par(mfrow=c(1, 2))
plot(st_geometry(mad0), main='Unprojected (WGS84)')
plot(st_geometry(mad0vert), main='Vertical')

# The effect is more noticable when plotting large areas,
# especially near the poles.
if (FALSE) {

library(geodata)
library(terra)

can <- gadm('CAN', level=0, path=getwd()) # outline of Canada

wktVert <- crsVertical(can)
wktLamb <- crsLambert(can)

canVert <- project(can, wktVert)
canLamb <- project(can, wktLamb)

par(mfrow=c(1, 3))
plot(can, main='WGS84')
plot(canVert, main='Vertical')
plot(canLamb, main='Lambert')

}

}
\seealso{
\code{\link{crsGet}}
}
