% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsVNS.r
\name{crsVNS}
\alias{crsVNS}
\title{Generate a "Vertical Near-Side" projection WKT2 string}
\usage{
crsVNS(x, alt = 35800)
}
\arguments{
\item{x}{Either an object of class \code{SpatRaster}, \code{SpatVector}, or \code{sf}, \emph{or} a numeric string with two values (longitude and latitude of the center of the projection), \emph{or} a two-column matrix/data frame with the centroid of the projection.}

\item{alt}{Altitude in meters of the viewpoint in km. The default (35800 km) is geosynchronous orbit.}
}
\description{
This function generates a WKT2 (well-known text) coordinate reference string for a "vertical near-side" projection of a geographic object or coordinate pair. When plotted, such an object will appear as if it would from a satellite in geosyncronous orbit directly above the centroid of the region of interest (appropriate for showing about 1/3 of the Earth).
}
\examples{
library(sf)

# Madagascar
data(mad0)

alb <- crsAlbers(mad0)
lamb <- crsLambert(mad0)
vert <- crsVertical(mad0)

madAlb <- st_transform(mad0, alb)
madLamb <- st_transform(mad0, lamb)
madVert <- st_transform(mad0, vert)

par(mfrow=c(2, 2))
plot(st_geometry(mad0), main='Unprojected (WGS84)')
plot(st_geometry(madAlb), main='Albers')
plot(st_geometry(madLamb), main='Lambert')
plot(st_geometry(madVert), main='Vertical')

# The effect is more noticable when plotting large areas,
# especially if they lie near the poles.
if (FALSE) {

library(geodata)
library(terra)

can <- gadm('CAN', level=0, path=tempdir()) # outline of Canada

alb <- crsAlbers(can)
lamb <- crsLambert(can)
vert <- crsVertical(can)

canAlb <- project(can, alb)
canLamb <- project(can, lamb)
canVert <- project(can, vert)

par(mfrow=c(2, 2))
plot(can, main='Unprojected (WGS84)')
plot(canAlb, main='Albers')
plot(canLamb, main='Lambert')
plot(canVert, main='Vertical')
	
}
}
\seealso{
\code{\link{crsGet}}, \code{\link{crsAlbers}}, \code{\link{crsLambert}}
}
