% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad1.r
\docType{data}
\name{mad1}
\alias{mad1}
\title{Madagascar spatial object}
\format{
An object of class \code{sf}.
}
\source{
\href{http://gadm.org}{GADM}
}
\usage{
data(mad1, package='enmSdmX')
}
\description{
Outlines of regions ("Faritra") of Madagascar from GADM. The geometry has been simplified from the version available in GADM, so pleased do not use this for "official" analyses.
}
\examples{
data(mad1)
mad1
plot(mad1['NAME_2'], main='Malagasy Faritra')
}
\keyword{Madagascar}
