% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCount.r
\name{pointCount}
\alias{pointCount}
\title{Number of points in a "spatial points" object}
\usage{
pointCount(x, byFeature = FALSE)
}
\description{
Returns the number of points in a \code{sf} or \code{SpatVector} object. This is typically done using either \code{length(x)} or \code{nrow(x)}, depending on whether the object in question has rows or not. This function helps in ambiguous cases, so users need not care if \code{nrow} or \code{length} is needed.
}
