% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad0.r
\docType{data}
\name{mad0}
\alias{mad0}
\title{Madagascar spatial object}
\format{
An object of class \code{sf}.
}
\source{
\href{http://www.gadm.org}{GADM}
}
\usage{
data(mad0, package='enmSdmX')
}
\description{
Outline of Madagascar from GADM.  The geometry has been simplified from the version available in GADM, so pleased do not use this for "official" analyses.
}
\examples{
data(mad0)
mad0
plot(mad0[1], main='Madagascar')
}
\keyword{Madagascar}
